requirements.txt
Django==2.1.1
mysqlclient==1.3.13
Pillow==5.2.0
pytz==2018.5
djangorestframework==3.7.3